message ingestion function modified

+++ src/main/java/com/example/notificationservice/constants/Constants.java
@@ -3,12 +3,12 @@ package com.example.notificationservice.constants;
 public class Constants {
 
     // Success Codes
-    public static final int INGESTION_SUCCESSFUL = 0;
-    public static final int UPDATE_SUCCESSFUL = 0;
+    public static final long INGESTION_SUCCESSFUL = 0;
+    public static final long UPDATE_SUCCESSFUL = 0;
 
     // Failure Codes
-    public static final int PHONE_NUMBER_MANDATORY = -1;
-    public static final int MESSAGE_WITH_ID_NOT_FOUND = -2;
+    public static final long PHONE_NUMBER_MANDATORY = -1;
+    public static final long MESSAGE_WITH_ID_NOT_FOUND = -2;
 
 
 }
